refactor: center pie chart
