+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from "react-router-dom";
 import PeopleTable from "./pages/PeopleTable";
 import NoMatch from "./pages/NoMatch/NoMatch";
-import PieChart from "./pages/PieChart";
+import PieChart from "./pages/PieChart/PieChart";
 import Header from "./layout/Header/Header";
 
 function App() {
