+++ src/pages/PieChart/PieChart.css
@@ -1,0 +1,6 @@
+.pie_chart {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
